[MTDSA-4632] backend allowances, deductions & reliefs alignment (#69)

+++ app/v1/models/response/getCalculation/allowancesAndDeductions/detail/ResidentialFinanceCosts.scala
@@ -18,10 +18,10 @@ package v1.models.response.getCalculation.allowancesAndDeductions.detail
 
 import play.api.libs.json._
 
-case class ResidentialFinanceCosts (amountClaimed: BigInt,
-                                    allowableAmount: Option[BigInt],
+case class ResidentialFinanceCosts (amountClaimed: BigDecimal,
+                                    allowableAmount: Option[BigDecimal],
                                     rate: BigDecimal,
-                                    propertyFinanceRelief: BigInt)
+                                    propertyFinanceRelief: BigDecimal)
 
 object ResidentialFinanceCosts {
   implicit val format: OFormat[ResidentialFinanceCosts] = Json.format[ResidentialFinanceCosts]
